+++ frontend/src/store/getters.js
@@ -474,7 +474,6 @@ export const getters = {
         video: true,
         image: true,
         popup: true,
-        highQuality: false
       },
       disableSettings: true,
       disableQuickToggles: false,
@@ -581,7 +580,6 @@ export const getters = {
         office: state.user?.preview?.office ?? true,
         folder: state.user?.preview?.folder ?? true,
         popup: state.user?.preview?.popup ?? true,
-        highQuality: state.user?.preview?.highQuality ?? false,
         motionVideoPreview: state.user?.preview?.motionVideoPreview ?? false,
         disableHideSidebar: state.user?.preview?.disableHideSidebar ?? false,
         autoplayMedia: state.user?.preview?.autoplayMedia ?? false,
@@ -596,7 +594,6 @@ export const getters = {
       office: state.user?.preview?.office ?? true,
       folder: state.user?.preview?.folder ?? true,
       popup: state.user?.preview?.popup ?? true,
-      highQuality: state.user?.preview?.highQuality ?? false,
       motionVideoPreview: state.user?.preview?.motionVideoPreview ?? false,
       disableHideSidebar: state.user?.preview?.disableHideSidebar ?? false,
       autoplayMedia: state.user?.preview?.autoplayMedia ?? false,