+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file. For commit 
 ## v1.3.0-beta
 
  **New Features**:
- - Archive/Unarchive actions in UI
+ - Archive/Unarchive actions in UI (#1252) (#335) (#1569) 
    - new api to archive/unarchive files on the server
    - requires `create` user permissions
    - archiving actions respect `server.maxArchiveSize`
- - Added share icon to items that are shared (requires share permission to see)
+ - Added share icon to items that are shared (#1420)
  - LDAP login support (#591)
+ - Enhanced thumbnail and item previews
+   - Added ability to show "motion preview" for folders with multiple child items that have previews. cycles through the first 4 images.
+   - support for reading embedded images from raw image or heic/heif files (#215)
+   - reorganized and simplified thumbnail settings in profile settings (#1968)
+   - removed `highQuality` thumbnail option which only affected gallery view. Now its always enabled.
+   - improved caching for unsupported images, the same file won't be attempted again with the same modtime.
+   - supports 3d model previews.
  - FileWatcher also supports watching directories
  - Support previews for 3D model files (STL, OBJ, 3MF, etc.) (#1273)
    - supported formats via threejs: GLTF, GLB, OBJ, STL, PLY, DAE (Collada), 3MF, 3DS, USDZ, USD, USDA, USDC, AMF, VRML, WRL, VTK, VTP, PCD, XYZ, VOX, KMZ
@@ -23,7 +30,7 @@ All notable changes to this project will be documented in this file. For commit 
    - Clicking outside of prompts no longer automatically closes them.
  - add webdav support (#209) -- thanks to @reddac for #1764
    - see docs on how to use
-   - requires "minimal" api token as password
+   - requires api an un-customized api token as password
    - respects access rules
    - requires download permission to view and modify/create/delete permission to modify.